requirements.txt
svn
pygithub
zipfiles